jimm/requirements.txt
alabaster==0.7.12
Babel==2.9.1
bandit==1.7.1
black==21.10b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
crypto==1.4.1
cryptography==36.0.0
cx-Freeze==6.8.2
docutils==0.17.1
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.4.0
ldapdomaindump==0.9.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
Naked==0.1.31
packaging==21.3
pathspec==0.9.0
pbr==5.7.0
pkginfo==1.8.2
platformdirs==2.4.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.12.0
pycryptodomex==3.11.0
Pygments==2.10.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytz==2021.3
PyYAML==6.0
readme-renderer==31.0
regex==2021.11.10
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.26
stevedore==3.5.0
tabulate==0.8.9
tomli==1.2.2
tqdm==4.62.3
twine==3.7.1
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
